requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
pytest-django
psycopg
drf-spectacular
hypothesis